added logger & git pkg
