+++ cmd/install.go
@@ -7,7 +7,8 @@ import (
 	"github.com/kyokomi/emoji"
 	"github.com/microsoft/fabrikate/core"
 	"github.com/microsoft/fabrikate/generators"
-	"github.com/microsoft/fabrikate/logger"
+	"github.com/microsoft/fabrikate/internal/git"
+	"github.com/microsoft/fabrikate/internal/logger"
 	"github.com/spf13/cobra"
 )
 
@@ -45,8 +46,8 @@ func Install(path string) (err error) {
 			return err
 		} else if len(accessTokens) > 0 {
 			for repo, token := range accessTokens {
-				if _, exists := core.GitAccessTokens.Get(repo); !exists {
-					core.GitAccessTokens.Set(repo, token)
+				if _, exists := git.GitAccessTokens.Get(repo); !exists {
+					git.GitAccessTokens.Set(repo, token)
 				}
 			}
 		}
