+++ generators/static.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/kyokomi/emoji"
 	"github.com/microsoft/fabrikate/core"
-	"github.com/microsoft/fabrikate/logger"
+	"github.com/microsoft/fabrikate/internal/logger"
 )
 
 // StaticGenerator uses a static directory of resource manifests to create a rolled up multi-part manifest.
